@@ -18,7 +18,6 @@ import (
 	"go/scanner"
 	"go/token"
 	"path/filepath"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -35,45 +34,19 @@ type cgoPackage struct {
 	packageDir      string // full path to the package to process
 	fset            *token.FileSet
 	tokenFiles      map[string]*token.File
-	missingSymbols  map[string]struct{}
-	constants       map[string]constantInfo
-	functions       map[string]*functionInfo
-	globals         map[string]globalInfo
-	typedefs        map[string]*typedefInfo
-	elaboratedTypes map[string]*elaboratedTypeInfo
-	enums           map[string]enumInfo
-	anonStructNum   int
+	definedGlobally map[string]ast.Node
+	anonDecls       map[interface{}]string
 	cflags          []string // CFlags from #cgo lines
 	ldflags         []string // LDFlags from #cgo lines
 	visitedFiles    map[string][]byte
 }
 
-// constantInfo stores some information about a CGo constant found by libclang
-// and declared in the Go AST.
-type constantInfo struct {
-	expr ast.Expr
-	pos  token.Pos
-}
-
-// functionInfo stores some information about a CGo function found by libclang
-// and declared in the AST.
-type functionInfo struct {
-	args     []paramInfo
-	results  *ast.FieldList
-	pos      token.Pos
-	variadic bool
-}
-
-// paramInfo is a parameter of a CGo function (see functionInfo).
-type paramInfo struct {
-	name     string
-	typeExpr ast.Expr
-}
-
-// typedefInfo contains information about a single typedef in C.
-type typedefInfo struct {
-	typeExpr ast.Expr
-	pos      token.Pos
+// cgoFile holds information only for a single Go file (with one or more
+// `import "C"` statements).
+type cgoFile struct {
+	*cgoPackage
+	defined map[string]ast.Node
+	names   map[string]clangCursor
 }
 
 // elaboratedTypeInfo contains some information about an elaborated type
@@ -97,18 +70,6 @@ type bitfieldInfo struct {
 	endBit   int64 // may be 0 meaning "until the end of the field"
 }
 
-// enumInfo contains information about an enum in the C.
-type enumInfo struct {
-	typeExpr ast.Expr
-	pos      token.Pos
-}
-
-// globalInfo contains information about a declared global variable in C.
-type globalInfo struct {
-	typeExpr ast.Expr
-	pos      token.Pos
-}
-
 // cgoAliases list type aliases between Go and C, for types that are equivalent
 // in both languages. See addTypeAliases.
 var cgoAliases = map[string]string{
@@ -125,24 +86,24 @@ var cgoAliases = map[string]string{
 
 // builtinAliases are handled specially because they only exist on the Go side
 // of CGo, not on the CGo side (they're prefixed with "_Cgo_" there).
-var builtinAliases = map[string]struct{}{
-	"char":      {},
-	"schar":     {},
-	"uchar":     {},
-	"short":     {},
-	"ushort":    {},
-	"int":       {},
-	"uint":      {},
-	"long":      {},
-	"ulong":     {},
-	"longlong":  {},
-	"ulonglong": {},
+var builtinAliases = []string{
+	"char",
+	"schar",
+	"uchar",
+	"short",
+	"ushort",
+	"int",
+	"uint",
+	"long",
+	"ulong",
+	"longlong",
+	"ulonglong",
 }
 
-// cgoTypes lists some C types with ambiguous sizes that must be retrieved
-// somehow from C. This is done by adding some typedefs to get the size of each
-// type.
-const cgoTypes = `
+// builtinAliasTypedefs lists some C types with ambiguous sizes that must be
+// retrieved somehow from C. This is done by adding some typedefs to get the
+// size of each type.
+const builtinAliasTypedefs = `
 # 1 "<cgo>"
 typedef char                _Cgo_char;
 typedef signed char         _Cgo_schar;
@@ -202,13 +163,8 @@ func Process(files []*ast.File, dir string, fset *token.FileSet, cflags []string
 		currentDir:      dir,
 		fset:            fset,
 		tokenFiles:      map[string]*token.File{},
-		missingSymbols:  map[string]struct{}{},
-		constants:       map[string]constantInfo{},
-		functions:       map[string]*functionInfo{},
-		globals:         map[string]globalInfo{},
-		typedefs:        map[string]*typedefInfo{},
-		elaboratedTypes: map[string]*elaboratedTypeInfo{},
-		enums:           map[string]enumInfo{},
+		definedGlobally: map[string]ast.Node{},
+		anonDecls:       map[interface{}]string{},
 		visitedFiles:    map[string][]byte{},
 	}
 
@@ -238,7 +194,7 @@ func Process(files []*ast.File, dir string, fset *token.FileSet, cflags []string
 		// This is always a bug in the cgo package.
 		panic("unexpected error: " + err.Error())
 	}
-	// If the Comments field is not set to nil, the fmt package will get
+	// If the Comments field is not set to nil, the go/format package will get
 	// confused about where comments should go.
 	p.generated.Comments = nil
 	// Adjust some of the functions in there.
@@ -254,15 +210,10 @@ func Process(files []*ast.File, dir string, fset *token.FileSet, cflags []string
 		}
 	}
 	// Patch some types, for example *C.char in C.CString.
-	astutil.Apply(p.generated, p.walker, nil)
-
-	// Find all C.* symbols.
-	for _, f := range files {
-		astutil.Apply(f, p.findMissingCGoNames, nil)
-	}
-	for name := range builtinAliases {
-		p.missingSymbols["_Cgo_"+name] = struct{}{}
-	}
+	cf := p.newCGoFile()
+	astutil.Apply(p.generated, func(cursor *astutil.Cursor) bool {
+		return cf.walker(cursor, nil)
+	}, nil)
 
 	// Find `import "C"` C fragments in the file.
 	cgoHeaders := make([]string, len(files)) // combined CGo header fragment for each file
@@ -337,44 +288,47 @@ func Process(files []*ast.File, dir string, fset *token.FileSet, cflags []string
 		cflagsForCGo = append(cflagsForCGo, "-isystem", clangHeaders)
 	}
 
+	// Retrieve types such as C.int, C.longlong, etc from C.
+	p.newCGoFile().readNames(builtinAliasTypedefs, cflagsForCGo, "", func(names map[string]clangCursor) {
+		gen := &ast.GenDecl{
+			TokPos: token.NoPos,
+			Tok:    token.TYPE,
+		}
+		for _, name := range builtinAliases {
+			typeSpec := p.getIntegerType("C."+name, names["_Cgo_"+name])
+			gen.Specs = append(gen.Specs, typeSpec)
+		}
+		p.generated.Decls = append(p.generated.Decls, gen)
+	})
+
 	// Process CGo imports for each file.
 	for i, f := range files {
-		p.parseFragment(cgoHeaders[i]+cgoTypes, cflagsForCGo, filepath.Base(fset.File(f.Pos()).Name()))
-	}
-
-	// Declare functions found by libclang.
-	p.addFuncDecls()
-
-	// Declare stub function pointer values found by libclang.
-	p.addFuncPtrDecls()
-
-	// Declare globals found by libclang.
-	p.addConstDecls()
-
-	// Declare globals found by libclang.
-	p.addVarDecls()
-
-	// Forward C types to Go types (like C.uint32_t -> uint32).
-	p.addTypeAliases()
-
-	// Add type declarations for C types, declared using typedef in C.
-	p.addTypedefs()
-
-	// Add elaborated types for C structs and unions.
-	p.addElaboratedTypes()
-
-	// Add enum types and enum constants for C enums.
-	p.addEnumTypes()
-
-	// Patch the AST to use the declared types and functions.
-	for _, f := range files {
-		astutil.Apply(f, p.walker, nil)
+		cf := p.newCGoFile()
+		cf.readNames(cgoHeaders[i], cflagsForCGo, filepath.Base(fset.File(f.Pos()).Name()), func(names map[string]clangCursor) {
+			for _, name := range builtinAliases {
+				// Names such as C.int should not be obtained from C.
+				// This works around an issue in picolibc that has `#define int`
+				// in a header file.
+				delete(names, name)
+			}
+			astutil.Apply(f, func(cursor *astutil.Cursor) bool {
+				return cf.walker(cursor, names)
+			}, nil)
+		})
 	}
 
 	// Print the newly generated in-memory AST, for debugging.
 	//ast.Print(fset, p.generated)
 
 	return p.generated, cgoHeaders, p.cflags, p.ldflags, p.visitedFiles, p.errors
+}
+
+func (p *cgoPackage) newCGoFile() *cgoFile {
+	return &cgoFile{
+		cgoPackage: p,
+		defined:    make(map[string]ast.Node),
+		names:      make(map[string]clangCursor),
+	}
 }
 
 // makePathsAbsolute converts some common path compiler flags (-I, -L) from
@@ -482,365 +436,6 @@ func (p *cgoPackage) parseCGoPreprocessorLines(text string, pos token.Pos) strin
 		}
 	}
 	return text
-}
-
-// addFuncDecls adds the C function declarations found by libclang in the
-// comment above the `import "C"` statement.
-func (p *cgoPackage) addFuncDecls() {
-	names := make([]string, 0, len(p.functions))
-	for name := range p.functions {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		fn := p.functions[name]
-		obj := &ast.Object{
-			Kind: ast.Fun,
-			Name: "C." + name,
-		}
-		args := make([]*ast.Field, len(fn.args))
-		decl := &ast.FuncDecl{
-			Doc: &ast.CommentGroup{
-				List: []*ast.Comment{
-					{
-						Slash: fn.pos - 1,
-						Text:  "//export " + name,
-					},
-				},
-			},
-			Name: &ast.Ident{
-				NamePos: fn.pos,
-				Name:    "C." + name,
-				Obj:     obj,
-			},
-			Type: &ast.FuncType{
-				Func: fn.pos,
-				Params: &ast.FieldList{
-					Opening: fn.pos,
-					List:    args,
-					Closing: fn.pos,
-				},
-				Results: fn.results,
-			},
-		}
-		if fn.variadic {
-			decl.Doc.List = append(decl.Doc.List, &ast.Comment{
-				Slash: fn.pos - 1,
-				Text:  "//go:variadic",
-			})
-		}
-		obj.Decl = decl
-		for i, arg := range fn.args {
-			args[i] = &ast.Field{
-				Names: []*ast.Ident{
-					{
-						NamePos: fn.pos,
-						Name:    arg.name,
-						Obj: &ast.Object{
-							Kind: ast.Var,
-							Name: arg.name,
-							Decl: decl,
-						},
-					},
-				},
-				Type: arg.typeExpr,
-			}
-		}
-		p.generated.Decls = append(p.generated.Decls, decl)
-	}
-}
-
-// addFuncPtrDecls creates stub declarations of function pointer values. These
-// values will later be replaced with the real values in the compiler.
-// It adds code like the following to the AST:
-//
-//     var (
-//         C.add unsafe.Pointer
-//         C.mul unsafe.Pointer
-//         // ...
-//     )
-func (p *cgoPackage) addFuncPtrDecls() {
-	if len(p.functions) == 0 {
-		return
-	}
-	names := make([]string, 0, len(p.functions))
-	for name := range p.functions {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		gen := &ast.GenDecl{
-			TokPos: token.NoPos,
-			Tok:    token.VAR,
-			Lparen: token.NoPos,
-			Rparen: token.NoPos,
-		}
-		fn := p.functions[name]
-		obj := &ast.Object{
-			Kind: ast.Typ,
-			Name: "C." + name + "$funcaddr",
-		}
-		valueSpec := &ast.ValueSpec{
-			Names: []*ast.Ident{{
-				NamePos: fn.pos,
-				Name:    "C." + name + "$funcaddr",
-				Obj:     obj,
-			}},
-			Type: &ast.SelectorExpr{
-				X: &ast.Ident{
-					NamePos: fn.pos,
-					Name:    "unsafe",
-				},
-				Sel: &ast.Ident{
-					NamePos: fn.pos,
-					Name:    "Pointer",
-				},
-			},
-		}
-		obj.Decl = valueSpec
-		gen.Specs = append(gen.Specs, valueSpec)
-		p.generated.Decls = append(p.generated.Decls, gen)
-	}
-}
-
-// addConstDecls declares external C constants in the Go source.
-// It adds code like the following to the AST:
-//
-//     const C.CONST_INT = 5
-//     const C.CONST_FLOAT = 5.8
-//     // ...
-func (p *cgoPackage) addConstDecls() {
-	if len(p.constants) == 0 {
-		return
-	}
-	names := make([]string, 0, len(p.constants))
-	for name := range p.constants {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		gen := &ast.GenDecl{
-			TokPos: token.NoPos,
-			Tok:    token.CONST,
-			Lparen: token.NoPos,
-			Rparen: token.NoPos,
-		}
-		constVal := p.constants[name]
-		obj := &ast.Object{
-			Kind: ast.Con,
-			Name: "C." + name,
-		}
-		valueSpec := &ast.ValueSpec{
-			Names: []*ast.Ident{{
-				NamePos: constVal.pos,
-				Name:    "C." + name,
-				Obj:     obj,
-			}},
-			Values: []ast.Expr{constVal.expr},
-		}
-		obj.Decl = valueSpec
-		gen.Specs = append(gen.Specs, valueSpec)
-		p.generated.Decls = append(p.generated.Decls, gen)
-	}
-}
-
-// addVarDecls declares external C globals in the Go source.
-// It adds code like the following to the AST:
-//
-//     var C.globalInt  int
-//     var C.globalBool bool
-//     // ...
-func (p *cgoPackage) addVarDecls() {
-	if len(p.globals) == 0 {
-		return
-	}
-	names := make([]string, 0, len(p.globals))
-	for name := range p.globals {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		global := p.globals[name]
-		gen := &ast.GenDecl{
-			TokPos: global.pos,
-			Tok:    token.VAR,
-			Lparen: token.NoPos,
-			Rparen: token.NoPos,
-			Doc: &ast.CommentGroup{
-				List: []*ast.Comment{
-					{
-						Slash: global.pos - 1,
-						Text:  "//go:extern " + name,
-					},
-				},
-			},
-		}
-		obj := &ast.Object{
-			Kind: ast.Var,
-			Name: "C." + name,
-		}
-		valueSpec := &ast.ValueSpec{
-			Names: []*ast.Ident{{
-				NamePos: global.pos,
-				Name:    "C." + name,
-				Obj:     obj,
-			}},
-			Type: global.typeExpr,
-		}
-		obj.Decl = valueSpec
-		gen.Specs = append(gen.Specs, valueSpec)
-		p.generated.Decls = append(p.generated.Decls, gen)
-	}
-}
-
-// addTypeAliases aliases some built-in Go types with their equivalent C types.
-// It adds code like the following to the AST:
-//
-//     type C.int8_t  = int8
-//     type C.int16_t = int16
-//     // ...
-func (p *cgoPackage) addTypeAliases() {
-	aliasKeys := make([]string, 0, len(cgoAliases))
-	for key := range cgoAliases {
-		aliasKeys = append(aliasKeys, key)
-	}
-	sort.Strings(aliasKeys)
-	for _, typeName := range aliasKeys {
-		gen := &ast.GenDecl{
-			TokPos: token.NoPos,
-			Tok:    token.TYPE,
-			Lparen: token.NoPos,
-			Rparen: token.NoPos,
-		}
-		goTypeName := cgoAliases[typeName]
-		obj := &ast.Object{
-			Kind: ast.Typ,
-			Name: typeName,
-		}
-		typeSpec := &ast.TypeSpec{
-			Name: &ast.Ident{
-				NamePos: token.NoPos,
-				Name:    typeName,
-				Obj:     obj,
-			},
-			Assign: p.generatedPos,
-			Type: &ast.Ident{
-				NamePos: token.NoPos,
-				Name:    goTypeName,
-			},
-		}
-		obj.Decl = typeSpec
-		gen.Specs = append(gen.Specs, typeSpec)
-		p.generated.Decls = append(p.generated.Decls, gen)
-	}
-}
-
-func (p *cgoPackage) addTypedefs() {
-	if len(p.typedefs) == 0 {
-		return
-	}
-	names := make([]string, 0, len(p.typedefs))
-	for name := range p.typedefs {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		gen := &ast.GenDecl{
-			TokPos: token.NoPos,
-			Tok:    token.TYPE,
-		}
-		typedef := p.typedefs[name]
-		typeName := "C." + name
-		isAlias := true
-		if strings.HasPrefix(name, "_Cgo_") {
-			typeName = "C." + name[len("_Cgo_"):]
-			isAlias = false // C.short etc. should not be aliased to the equivalent Go type (not portable)
-		}
-		if _, ok := cgoAliases[typeName]; ok {
-			// This is a type that also exists in Go (defined in stdint.h).
-			continue
-		}
-		obj := &ast.Object{
-			Kind: ast.Typ,
-			Name: typeName,
-		}
-		typeSpec := &ast.TypeSpec{
-			Name: &ast.Ident{
-				NamePos: typedef.pos,
-				Name:    typeName,
-				Obj:     obj,
-			},
-			Type: typedef.typeExpr,
-		}
-		if isAlias {
-			typeSpec.Assign = typedef.pos
-		}
-		obj.Decl = typeSpec
-		gen.Specs = append(gen.Specs, typeSpec)
-		p.generated.Decls = append(p.generated.Decls, gen)
-	}
-}
-
-// addElaboratedTypes adds C elaborated types as aliases. These are the "struct
-// foo" or "union foo" types, often used in a typedef.
-//
-// See also:
-// https://en.cppreference.com/w/cpp/language/elaborated_type_specifier
-func (p *cgoPackage) addElaboratedTypes() {
-	if len(p.elaboratedTypes) == 0 {
-		return
-	}
-	names := make([]string, 0, len(p.elaboratedTypes))
-	for name := range p.elaboratedTypes {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		gen := &ast.GenDecl{
-			TokPos: token.NoPos,
-			Tok:    token.TYPE,
-		}
-		typ := p.elaboratedTypes[name]
-		typeName := "C." + name
-		obj := &ast.Object{
-			Kind: ast.Typ,
-			Name: typeName,
-		}
-		typeExpr := typ.typeExpr
-		if typ.unionSize != 0 {
-			// Create getters/setters.
-			for _, field := range typ.typeExpr.Fields.List {
-				if len(field.Names) != 1 {
-					p.addError(typ.pos, fmt.Sprintf("union must have field with a single name, it has %d names", len(field.Names)))
-					continue
-				}
-				p.createUnionAccessor(field, typeName)
-			}
-			// Convert to a single-field struct type.
-			typeExpr = p.makeUnionField(typ)
-			if typeExpr == nil {
-				// There was an error, that was already added to the list of
-				// errors.
-				continue
-			}
-		}
-		typeSpec := &ast.TypeSpec{
-			Name: &ast.Ident{
-				NamePos: typ.pos,
-				Name:    typeName,
-				Obj:     obj,
-			},
-			Type: typeExpr,
-		}
-		obj.Decl = typeSpec
-		gen.Specs = append(gen.Specs, typeSpec)
-		// If this struct has bitfields, create getters for them.
-		for _, bitfield := range typ.bitfields {
-			p.createBitfieldGetter(bitfield, typeName)
-			p.createBitfieldSetter(bitfield, typeName)
-		}
-		p.generated.Decls = append(p.generated.Decls, gen)
-	}
 }
 
 // makeUnionField creates a new struct from an existing *elaboratedTypeInfo,
@@ -1304,80 +899,319 @@ func (p *cgoPackage) createBitfieldSetter(bitfield bitfieldInfo, typeName string
 	p.generated.Decls = append(p.generated.Decls, setter)
 }
 
-// addEnumTypes adds C enums to the AST. For example, the following C code:
-//
-//     enum option {
-//         optionA,
-//         optionB = 5,
-//     };
-//
-// is translated to the following Go code equivalent:
-//
-//     type C.enum_option int32
-//
-// The constants are treated just like macros so are inserted into the AST by
-// addConstDecls.
-// See also: https://en.cppreference.com/w/c/language/enum
-func (p *cgoPackage) addEnumTypes() {
-	if len(p.enums) == 0 {
-		return
-	}
-	names := make([]string, 0, len(p.enums))
-	for name := range p.enums {
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	for _, name := range names {
-		gen := &ast.GenDecl{
-			TokPos: token.NoPos,
-			Tok:    token.TYPE,
+// isEquivalentAST returns whether the given two AST nodes are equivalent as far
+// as CGo is concerned. This is used to check that C types, globals, etc defined
+// in different CGo header snippets are actually the same type (and probably
+// even defined in the same header file, just in different translation units).
+func (p *cgoPackage) isEquivalentAST(a, b ast.Node) bool {
+	switch node := a.(type) {
+	case *ast.ArrayType:
+		b, ok := b.(*ast.ArrayType)
+		if !ok {
+			return false
 		}
-		typ := p.enums[name]
-		typeName := "C.enum_" + name
-		obj := &ast.Object{
-			Kind: ast.Typ,
-			Name: typeName,
+		if !p.isEquivalentAST(node.Len, b.Len) {
+			return false
 		}
-		typeSpec := &ast.TypeSpec{
-			Name: &ast.Ident{
-				NamePos: typ.pos,
-				Name:    typeName,
-				Obj:     obj,
-			},
-			Type: typ.typeExpr,
+		return p.isEquivalentAST(node.Elt, b.Elt)
+	case *ast.BasicLit:
+		b, ok := b.(*ast.BasicLit)
+		if !ok {
+			return false
 		}
-		obj.Decl = typeSpec
-		gen.Specs = append(gen.Specs, typeSpec)
-		p.generated.Decls = append(p.generated.Decls, gen)
+		// Note: this comparison is not correct in general ("1e2" equals "100"),
+		// but is correct for its use in the cgo package.
+		return node.Value == b.Value
+	case *ast.CommentGroup:
+		b, ok := b.(*ast.CommentGroup)
+		if !ok {
+			return false
+		}
+		if len(node.List) != len(b.List) {
+			return false
+		}
+		for i, c := range node.List {
+			if c.Text != b.List[i].Text {
+				return false
+			}
+		}
+		return true
+	case *ast.FieldList:
+		b, ok := b.(*ast.FieldList)
+		if !ok {
+			return false
+		}
+		if len(node.List) != len(b.List) {
+			return false
+		}
+		for i, f := range node.List {
+			if !p.isEquivalentAST(f, b.List[i]) {
+				return false
+			}
+		}
+		return true
+	case *ast.Field:
+		b, ok := b.(*ast.Field)
+		if !ok {
+			return false
+		}
+		if !p.isEquivalentAST(node.Type, b.Type) {
+			return false
+		}
+		if len(node.Names) != len(b.Names) {
+			return false
+		}
+		for i, name := range node.Names {
+			if name.Name != b.Names[i].Name {
+				return false
+			}
+		}
+		return true
+	case *ast.FuncDecl:
+		b, ok := b.(*ast.FuncDecl)
+		if !ok {
+			return false
+		}
+		if node.Name.Name != b.Name.Name {
+			return false
+		}
+		if node.Doc != b.Doc {
+			if !p.isEquivalentAST(node.Doc, b.Doc) {
+				return false
+			}
+		}
+		if node.Recv != b.Recv {
+			if !p.isEquivalentAST(node.Recv, b.Recv) {
+				return false
+			}
+		}
+		if !p.isEquivalentAST(node.Type.Params, b.Type.Params) {
+			return false
+		}
+		return p.isEquivalentAST(node.Type.Results, b.Type.Results)
+	case *ast.GenDecl:
+		b, ok := b.(*ast.GenDecl)
+		if !ok {
+			return false
+		}
+		if node.Doc != b.Doc {
+			if !p.isEquivalentAST(node.Doc, b.Doc) {
+				return false
+			}
+		}
+		if len(node.Specs) != len(b.Specs) {
+			return false
+		}
+		for i, s := range node.Specs {
+			if !p.isEquivalentAST(s, b.Specs[i]) {
+				return false
+			}
+		}
+		return true
+	case *ast.Ident:
+		b, ok := b.(*ast.Ident)
+		if !ok {
+			return false
+		}
+		return node.Name == b.Name
+	case *ast.SelectorExpr:
+		b, ok := b.(*ast.SelectorExpr)
+		if !ok {
+			return false
+		}
+		if !p.isEquivalentAST(node.Sel, b.Sel) {
+			return false
+		}
+		return p.isEquivalentAST(node.X, b.X)
+	case *ast.StarExpr:
+		b, ok := b.(*ast.StarExpr)
+		if !ok {
+			return false
+		}
+		return p.isEquivalentAST(node.X, b.X)
+	case *ast.StructType:
+		b, ok := b.(*ast.StructType)
+		if !ok {
+			return false
+		}
+		return p.isEquivalentAST(node.Fields, b.Fields)
+	case *ast.TypeSpec:
+		b, ok := b.(*ast.TypeSpec)
+		if !ok {
+			return false
+		}
+		if node.Name.Name != b.Name.Name {
+			return false
+		}
+		if node.Assign.IsValid() != b.Assign.IsValid() {
+			return false
+		}
+		return p.isEquivalentAST(node.Type, b.Type)
+	case *ast.ValueSpec:
+		b, ok := b.(*ast.ValueSpec)
+		if !ok {
+			return false
+		}
+		if len(node.Names) != len(b.Names) {
+			return false
+		}
+		for i, name := range node.Names {
+			if name.Name != b.Names[i].Name {
+				return false
+			}
+		}
+		if node.Type != b.Type && !p.isEquivalentAST(node.Type, b.Type) {
+			return false
+		}
+		if len(node.Values) != len(b.Values) {
+			return false
+		}
+		for i, value := range node.Values {
+			if !p.isEquivalentAST(value, b.Values[i]) {
+				return false
+			}
+		}
+		return true
+	case nil:
+		p.addError(token.NoPos, "internal error: AST node is nil")
+		return true
+	default:
+		p.addError(a.Pos(), fmt.Sprintf("internal error: unknown AST node: %T", a))
+		return true
 	}
 }
 
-// findMissingCGoNames traverses the AST and finds all C.something names. Only
-// these symbols are extracted from the parsed C AST and converted to the Go
-// equivalent.
-func (p *cgoPackage) findMissingCGoNames(cursor *astutil.Cursor) bool {
-	switch node := cursor.Node().(type) {
-	case *ast.SelectorExpr:
-		x, ok := node.X.(*ast.Ident)
-		if !ok {
-			return true
+// getPos returns node.Pos(), and tries to obtain a closely related position if
+// that fails.
+func getPos(node ast.Node) token.Pos {
+	pos := node.Pos()
+	if pos.IsValid() {
+		return pos
+	}
+	if decl, ok := node.(*ast.GenDecl); ok {
+		// *ast.GenDecl often doesn't have TokPos defined, so look at the first
+		// spec.
+		return getPos(decl.Specs[0])
+	}
+	return token.NoPos
+}
+
+// getUnnamedDeclName creates a name (with the given prefix) for the given C
+// declaration. This is used for structs, unions, and enums that are often
+// defined without a name and used in a typedef.
+func (p *cgoPackage) getUnnamedDeclName(prefix string, itf interface{}) string {
+	if name, ok := p.anonDecls[itf]; ok {
+		return name
+	}
+	name := prefix + strconv.Itoa(len(p.anonDecls))
+	p.anonDecls[itf] = name
+	return name
+}
+
+// getASTDeclName will declare the given C AST node (if not already defined) and
+// will return its name, in the form of C.foo.
+func (f *cgoFile) getASTDeclName(name string, found clangCursor, iscall bool) string {
+	// Some types are defined in stdint.h and map directly to a particular Go
+	// type.
+	if alias := cgoAliases["C."+name]; alias != "" {
+		return alias
+	}
+	node := f.getASTDeclNode(name, found, iscall)
+	if _, ok := node.(*ast.FuncDecl); ok && !iscall {
+		return "C." + name + "$funcaddr"
+	}
+	return "C." + name
+}
+
+// getASTDeclNode will declare the given C AST node (if not already defined) and
+// returns it.
+func (f *cgoFile) getASTDeclNode(name string, found clangCursor, iscall bool) ast.Node {
+	if node, ok := f.defined[name]; ok {
+		// Declaration was found in the current file, so return it immediately.
+		return node
+	}
+
+	if node, ok := f.definedGlobally[name]; ok {
+		// Declaration was previously created, but not for the current file. It
+		// may be different (because it comes from a different CGo snippet), so
+		// we need to check whether the AST for this definition is equivalent.
+		f.defined[name] = nil
+		newNode, _ := f.createASTNode(name, found)
+		if !f.isEquivalentAST(node, newNode) {
+			// It's not. Return a nice error with both locations.
+			// Original cgo reports an error like
+			//   cgo: inconsistent definitions for C.myint
+			// which is far less helpful.
+			f.addError(getPos(node), "defined previously at "+f.fset.Position(getPos(newNode)).String()+" with a different type")
 		}
-		if x.Name == "C" {
-			name := node.Sel.Name
-			if _, ok := builtinAliases[name]; ok {
-				name = "_Cgo_" + name
+		f.defined[name] = node
+		return node
+	}
+
+	// The declaration has no AST node. Create it now.
+	f.defined[name] = nil
+	node, elaboratedType := f.createASTNode(name, found)
+	f.defined[name] = node
+	f.definedGlobally[name] = node
+	switch node := node.(type) {
+	case *ast.FuncDecl:
+		f.generated.Decls = append(f.generated.Decls, node)
+		// Also add a declaration like the following:
+		//   var C.foo$funcaddr unsafe.Pointer
+		f.generated.Decls = append(f.generated.Decls, &ast.GenDecl{
+			Tok: token.VAR,
+			Specs: []ast.Spec{
+				&ast.ValueSpec{
+					Names: []*ast.Ident{{Name: "C." + name + "$funcaddr"}},
+					Type: &ast.SelectorExpr{
+						X:   &ast.Ident{Name: "unsafe"},
+						Sel: &ast.Ident{Name: "Pointer"},
+					},
+				},
+			},
+		})
+	case *ast.GenDecl:
+		f.generated.Decls = append(f.generated.Decls, node)
+	case *ast.TypeSpec:
+		f.generated.Decls = append(f.generated.Decls, &ast.GenDecl{
+			Tok:   token.TYPE,
+			Specs: []ast.Spec{node},
+		})
+	case nil:
+		// Node may be nil in case of an error. In that case, just don't add it
+		// as a declaration.
+	default:
+		panic("unexpected AST node")
+	}
+
+	// If this is a struct or union it may need bitfields or union accessor
+	// methods.
+	if elaboratedType != nil {
+		// Add struct bitfields.
+		for _, bitfield := range elaboratedType.bitfields {
+			f.createBitfieldGetter(bitfield, "C."+name)
+			f.createBitfieldSetter(bitfield, "C."+name)
+		}
+		if elaboratedType.unionSize != 0 {
+			// Create union getters/setters.
+			for _, field := range elaboratedType.typeExpr.Fields.List {
+				if len(field.Names) != 1 {
+					f.addError(elaboratedType.pos, fmt.Sprintf("union must have field with a single name, it has %d names", len(field.Names)))
+					continue
+				}
+				f.createUnionAccessor(field, "C."+name)
 			}
-			p.missingSymbols[name] = struct{}{}
 		}
 	}
-	return true
+
+	return node
 }
 
 // walker replaces all "C".<something> expressions to literal "C.<something>"
 // expressions. Such expressions are impossible to write in Go (a dot cannot be
 // used in the middle of a name) so in practice all C identifiers live in a
 // separate namespace (no _Cgo_ hacks like in gc).
-func (p *cgoPackage) walker(cursor *astutil.Cursor) bool {
+func (f *cgoFile) walker(cursor *astutil.Cursor, names map[string]clangCursor) bool {
 	switch node := cursor.Node().(type) {
 	case *ast.CallExpr:
 		fun, ok := node.Fun.(*ast.SelectorExpr)
@@ -1388,10 +1222,10 @@ func (p *cgoPackage) walker(cursor *astutil.Cursor) bool {
 		if !ok {
 			return true
 		}
-		if _, ok := p.functions[fun.Sel.Name]; ok && x.Name == "C" {
+		if found, ok := names[fun.Sel.Name]; ok && x.Name == "C" {
 			node.Fun = &ast.Ident{
 				NamePos: x.NamePos,
-				Name:    "C." + fun.Sel.Name,
+				Name:    f.getASTDeclName(fun.Sel.Name, found, true),
 			}
 		}
 	case *ast.SelectorExpr:
@@ -1401,8 +1235,8 @@ func (p *cgoPackage) walker(cursor *astutil.Cursor) bool {
 		}
 		if x.Name == "C" {
 			name := "C." + node.Sel.Name
-			if _, ok := p.functions[node.Sel.Name]; ok {
-				name += "$funcaddr"
+			if found, ok := names[node.Sel.Name]; ok {
+				name = f.getASTDeclName(node.Sel.Name, found, false)
 			}
 			cursor.Replace(&ast.Ident{
 				NamePos: x.NamePos,
