@@ -233,14 +233,6 @@ func setHeapEnd(newHeapEnd uintptr) {
 // This function can be called again when the heap size increases. The caller is
 // responsible for copying the metadata to the new location.
 func calculateHeapAddresses() {
-	if GOARCH == "wasm" {
-		// This is a workaround for a bug in wasm-ld: wasm-ld doesn't always
-		// align __heap_base and when this memory is shared through an API, it
-		// might result in unaligned memory. For details, see:
-		// https://reviews.llvm.org/D106499
-		// It should be removed once we switch to LLVM 13, where this is fixed.
-		heapStart = align(heapStart)
-	}
 	totalSize := heapEnd - heapStart
 
 	// Allocate some memory to keep 2 bits of information about every block.
@@ -310,6 +302,13 @@ func alloc(size uintptr, layout unsafe.Pointer) unsafe.Pointer {
 			index = 0
 			// Reset numFreeBlocks as allocations cannot wrap.
 			numFreeBlocks = 0
+			// In rare cases, the initial heap might be so small that there are
+			// no blocks at all. In this case, it's better to jump back to the
+			// start of the loop and try again, until the GC realizes there is
+			// no memory and grows the heap.
+			// This can sometimes happen on WebAssembly, where the initial heap
+			// is created by whatever is left on the last memory page.
+			continue
 		}
 
 		// Is the block we're looking at free?
