@@ -11,6 +11,8 @@ import (
 	"unsafe"
 )
 
+const gcAsserts = false // perform sanity checks
+
 // Ever-incrementing pointer: no memory is freed.
 var heapptr = heapStart
 
@@ -67,12 +69,7 @@ func SetFinalizer(obj interface{}, finalizer interface{}) {
 
 func initHeap() {
 	// preinit() may have moved heapStart; reset heapptr
-	ptr := heapStart
-	if GOARCH == "wasm" {
-		// llvm11 and llvm12 do not correctly align the heap on wasm
-		ptr = align(ptr)
-	}
-	heapptr = ptr
+	heapptr = heapStart
 }
 
 // setHeapEnd sets a new (larger) heapEnd pointer.
@@ -80,4 +77,8 @@ func setHeapEnd(newHeapEnd uintptr) {
 	// This "heap" is so simple that simply assigning a new value is good
 	// enough.
 	heapEnd = newHeapEnd
+}
+
+func markRoots(start, end uintptr) {
+	// dummy, so that markGlobals will compile
 }
