@@ -12,13 +12,22 @@ import (
 
 // The underlying hashmap structure for Go.
 type hashmap struct {
-	next       *hashmap       // hashmap after evacuate (for iterators)
 	buckets    unsafe.Pointer // pointer to array of buckets
 	count      uintptr
 	keySize    uint8 // maybe this can store the key type as well? E.g. keysize == 5 means string?
 	valueSize  uint8
 	bucketBits uint8
+	keyEqual   func(x, y unsafe.Pointer, n uintptr) bool
+	keyHash    func(key unsafe.Pointer, size uintptr) uint32
 }
+
+type hashmapAlgorithm uint8
+
+const (
+	hashmapAlgorithmBinary hashmapAlgorithm = iota
+	hashmapAlgorithmString
+	hashmapAlgorithmInterface
+)
 
 // A hashmap bucket. A bucket is a container of 8 key/value pairs: first the
 // following two entries, then the 8 keys, then the 8 values. This somewhat odd
@@ -32,9 +41,11 @@ type hashmapBucket struct {
 }
 
 type hashmapIterator struct {
-	bucketNumber uintptr
-	bucket       *hashmapBucket
-	bucketIndex  uint8
+	buckets      unsafe.Pointer // pointer to array of hashapBuckets
+	numBuckets   uintptr        // length of buckets array
+	bucketNumber uintptr        // current index into buckets array
+	bucket       *hashmapBucket // current bucket in chain
+	bucketIndex  uint8          // current index into bucket
 }
 
 // Get the topmost 8 bits of the hash, without using a special value (like 0).
@@ -48,7 +59,7 @@ func hashmapTopHash(hash uint32) uint8 {
 }
 
 // Create a new hashmap with the given keySize and valueSize.
-func hashmapMake(keySize, valueSize uint8, sizeHint uintptr) *hashmap {
+func hashmapMake(keySize, valueSize uint8, sizeHint uintptr, alg uint8) *hashmap {
 	numBuckets := sizeHint / 8
 	bucketBits := uint8(0)
 	for numBuckets != 0 {
@@ -57,12 +68,65 @@ func hashmapMake(keySize, valueSize uint8, sizeHint uintptr) *hashmap {
 	}
 	bucketBufSize := unsafe.Sizeof(hashmapBucket{}) + uintptr(keySize)*8 + uintptr(valueSize)*8
 	buckets := alloc(bucketBufSize*(1<<bucketBits), nil)
+
+	keyHash := hashmapKeyHashAlg(hashmapAlgorithm(alg))
+	keyEqual := hashmapKeyEqualAlg(hashmapAlgorithm(alg))
+
 	return &hashmap{
 		buckets:    buckets,
 		keySize:    keySize,
 		valueSize:  valueSize,
 		bucketBits: bucketBits,
+		keyEqual:   keyEqual,
+		keyHash:    keyHash,
 	}
+}
+
+func hashmapKeyEqualAlg(alg hashmapAlgorithm) func(x, y unsafe.Pointer, n uintptr) bool {
+	switch alg {
+	case hashmapAlgorithmBinary:
+		return memequal
+	case hashmapAlgorithmString:
+		return hashmapStringEqual
+	case hashmapAlgorithmInterface:
+		return hashmapInterfaceEqual
+	default:
+		// compiler bug :(
+		return nil
+	}
+}
+
+func hashmapKeyHashAlg(alg hashmapAlgorithm) func(key unsafe.Pointer, n uintptr) uint32 {
+	switch alg {
+	case hashmapAlgorithmBinary:
+		return hash32
+	case hashmapAlgorithmString:
+		return hashmapStringPtrHash
+	case hashmapAlgorithmInterface:
+		return hashmapInterfacePtrHash
+	default:
+		// compiler bug :(
+		return nil
+	}
+}
+
+func hashmapShouldGrow(m *hashmap) bool {
+	if m.bucketBits > uint8((unsafe.Sizeof(uintptr(0))*8)-3) {
+		// Over this limit, we're likely to overflow uintptrs during calculations
+		// or numbers of hash elements.   Don't allow any more growth.
+		// With 29 bits, this is 2^32 elements anyway.
+		return false
+	}
+
+	// "maximum" number of elements is 0.75 * buckets * elements per bucket
+	// to avoid overflow, this is calculated as
+	// max = 3 * (1/4 * buckets * elements per bucket)
+	//     = 3 * (buckets * (elements per bucket)/4)
+	//     = 3 * (buckets * (8/4)
+	//     = 3 * (buckets * 2)
+	//     = 6 * buckets
+	max := (uintptr(6) << m.bucketBits)
+	return m.count > max
 }
 
 // Return the number of entries in this hashmap, called from the len builtin.
@@ -83,13 +147,11 @@ func hashmapLenUnsafePointer(p unsafe.Pointer) int {
 
 // Set a specified key to a given value. Grow the map if necessary.
 //go:nobounds
-func hashmapSet(m *hashmap, key unsafe.Pointer, value unsafe.Pointer, hash uint32, keyEqual func(x, y unsafe.Pointer, n uintptr) bool) {
+func hashmapSet(m *hashmap, key unsafe.Pointer, value unsafe.Pointer, hash uint32) {
 	tophash := hashmapTopHash(hash)
 
-	if m.buckets == nil {
-		// No bucket was allocated yet, do so now.
-		m.buckets = unsafe.Pointer(hashmapInsertIntoNewBucket(m, key, value, tophash))
-		return
+	if hashmapShouldGrow(m) {
+		hashmapGrow(m)
 	}
 
 	numBuckets := uintptr(1) << m.bucketBits
@@ -118,7 +180,7 @@ func hashmapSet(m *hashmap, key unsafe.Pointer, value unsafe.Pointer, hash uint3
 			}
 			if bucket.tophash[i] == tophash {
 				// Could be an existing key that's the same.
-				if keyEqual(key, slotKey, uintptr(m.keySize)) {
+				if m.keyEqual(key, slotKey, uintptr(m.keySize)) {
 					// found same key, replace it
 					memcpy(slotValue, value, uintptr(m.valueSize))
 					return
@@ -158,9 +220,35 @@ func hashmapInsertIntoNewBucket(m *hashmap, key, value unsafe.Pointer, tophash u
 	return bucket
 }
 
+func hashmapGrow(m *hashmap) {
+
+	// clone map as empty
+	n := *m
+	n.count = 0
+
+	// allocate our new buckets twice as big
+	n.bucketBits = m.bucketBits + 1
+	numBuckets := uintptr(1) << n.bucketBits
+	bucketBufSize := unsafe.Sizeof(hashmapBucket{}) + uintptr(m.keySize)*8 + uintptr(m.valueSize)*8
+	n.buckets = alloc(bucketBufSize*numBuckets, nil)
+
+	// use a hashmap iterator to go through the old map
+	var it hashmapIterator
+
+	var key = alloc(uintptr(m.keySize), nil)
+	var value = alloc(uintptr(m.valueSize), nil)
+
+	for hashmapNext(m, &it, key, value) {
+		h := m.keyHash(key, uintptr(m.keySize))
+		hashmapSet(&n, key, value, h)
+	}
+
+	*m = n
+}
+
 // Get the value of a specified key, or zero the value if not found.
 //go:nobounds
-func hashmapGet(m *hashmap, key, value unsafe.Pointer, valueSize uintptr, hash uint32, keyEqual func(x, y unsafe.Pointer, n uintptr) bool) bool {
+func hashmapGet(m *hashmap, key, value unsafe.Pointer, valueSize uintptr, hash uint32) bool {
 	if m == nil {
 		// Getting a value out of a nil map is valid. From the spec:
 		// > if the map is nil or does not contain such an entry, a[x] is the
@@ -189,7 +277,7 @@ func hashmapGet(m *hashmap, key, value unsafe.Pointer, valueSize uintptr, hash u
 			slotValue := unsafe.Pointer(uintptr(unsafe.Pointer(bucket)) + slotValueOffset)
 			if bucket.tophash[i] == tophash {
 				// This could be the key we're looking for.
-				if keyEqual(key, slotKey, uintptr(m.keySize)) {
+				if m.keyEqual(key, slotKey, uintptr(m.keySize)) {
 					// Found the key, copy it.
 					memcpy(value, slotValue, uintptr(m.valueSize))
 					return true
@@ -207,7 +295,7 @@ func hashmapGet(m *hashmap, key, value unsafe.Pointer, valueSize uintptr, hash u
 // Delete a given key from the map. No-op when the key does not exist in the
 // map.
 //go:nobounds
-func hashmapDelete(m *hashmap, key unsafe.Pointer, hash uint32, keyEqual func(x, y unsafe.Pointer, n uintptr) bool) {
+func hashmapDelete(m *hashmap, key unsafe.Pointer, hash uint32) {
 	if m == nil {
 		// The delete builtin is defined even when the map is nil. From the spec:
 		// > If the map m is nil or the element m[k] does not exist, delete is a
@@ -233,7 +321,7 @@ func hashmapDelete(m *hashmap, key unsafe.Pointer, hash uint32, keyEqual func(x,
 			slotKey := unsafe.Pointer(uintptr(unsafe.Pointer(bucket)) + slotKeyOffset)
 			if bucket.tophash[i] == tophash {
 				// This could be the key we're looking for.
-				if keyEqual(key, slotKey, uintptr(m.keySize)) {
+				if m.keyEqual(key, slotKey, uintptr(m.keySize)) {
 					// Found the key, delete it.
 					bucket.tophash[i] = 0
 					m.count--
@@ -249,13 +337,16 @@ func hashmapDelete(m *hashmap, key unsafe.Pointer, hash uint32, keyEqual func(x,
 //go:nobounds
 func hashmapNext(m *hashmap, it *hashmapIterator, key, value unsafe.Pointer) bool {
 	if m == nil {
-		// Iterating over a nil slice appears to be allowed by the Go spec:
-		// https://groups.google.com/g/golang-nuts/c/gVgVLQU1FFE?pli=1
-		// https://play.golang.org/p/S8jxAMytKDB
+		// From the spec: If the map is nil, the number of iterations is 0.
 		return false
 	}
 
-	numBuckets := uintptr(1) << m.bucketBits
+	if it.buckets == nil {
+		// initialize iterator
+		it.buckets = m.buckets
+		it.numBuckets = uintptr(1) << m.bucketBits
+	}
+
 	for {
 		if it.bucketIndex >= 8 {
 			// end of bucket, move to the next in the chain
@@ -263,12 +354,12 @@ func hashmapNext(m *hashmap, it *hashmapIterator, key, value unsafe.Pointer) boo
 			it.bucket = it.bucket.next
 		}
 		if it.bucket == nil {
-			if it.bucketNumber >= numBuckets {
+			if it.bucketNumber >= it.numBuckets {
 				// went through all buckets
 				return false
 			}
 			bucketSize := unsafe.Sizeof(hashmapBucket{}) + uintptr(m.keySize)*8 + uintptr(m.valueSize)*8
-			bucketAddr := uintptr(m.buckets) + bucketSize*it.bucketNumber
+			bucketAddr := uintptr(it.buckets) + bucketSize*it.bucketNumber
 			it.bucket = (*hashmapBucket)(unsafe.Pointer(bucketAddr))
 			it.bucketNumber++ // next bucket
 		}
@@ -281,11 +372,29 @@ func hashmapNext(m *hashmap, it *hashmapIterator, key, value unsafe.Pointer) boo
 		bucketAddr := uintptr(unsafe.Pointer(it.bucket))
 		slotKeyOffset := unsafe.Sizeof(hashmapBucket{}) + uintptr(m.keySize)*uintptr(it.bucketIndex)
 		slotKey := unsafe.Pointer(bucketAddr + slotKeyOffset)
-		slotValueOffset := unsafe.Sizeof(hashmapBucket{}) + uintptr(m.keySize)*8 + uintptr(m.valueSize)*uintptr(it.bucketIndex)
-		slotValue := unsafe.Pointer(bucketAddr + slotValueOffset)
 		memcpy(key, slotKey, uintptr(m.keySize))
-		memcpy(value, slotValue, uintptr(m.valueSize))
-		it.bucketIndex++
+
+		if it.buckets == m.buckets {
+			// Our view of the buckets is the same as the parent map.
+			// Just copy the value we have
+			slotValueOffset := unsafe.Sizeof(hashmapBucket{}) + uintptr(m.keySize)*8 + uintptr(m.valueSize)*uintptr(it.bucketIndex)
+			slotValue := unsafe.Pointer(bucketAddr + slotValueOffset)
+			memcpy(value, slotValue, uintptr(m.valueSize))
+			it.bucketIndex++
+		} else {
+			it.bucketIndex++
+
+			// Our view of the buckets doesn't match the parent map.
+			// Look up the key in the new buckets and return that value if it exists
+			hash := m.keyHash(key, uintptr(m.keySize))
+			ok := hashmapGet(m, key, value, uintptr(m.valueSize), hash)
+			if !ok {
+				// doesn't exist in parent map; try next key
+				continue
+			}
+
+			// All good.
+		}
 
 		return true
 	}
@@ -296,7 +405,7 @@ func hashmapNext(m *hashmap, it *hashmapIterator, key, value unsafe.Pointer) boo
 func hashmapBinarySet(m *hashmap, key, value unsafe.Pointer) {
 	// TODO: detect nil map here and throw a better panic message?
 	hash := hash32(key, uintptr(m.keySize))
-	hashmapSet(m, key, value, hash, memequal)
+	hashmapSet(m, key, value, hash)
 }
 
 func hashmapBinaryGet(m *hashmap, key, value unsafe.Pointer, valueSize uintptr) bool {
@@ -305,7 +414,7 @@ func hashmapBinaryGet(m *hashmap, key, value unsafe.Pointer, valueSize uintptr) 
 		return false
 	}
 	hash := hash32(key, uintptr(m.keySize))
-	return hashmapGet(m, key, value, valueSize, hash, memequal)
+	return hashmapGet(m, key, value, valueSize, hash)
 }
 
 func hashmapBinaryDelete(m *hashmap, key unsafe.Pointer) {
@@ -313,7 +422,7 @@ func hashmapBinaryDelete(m *hashmap, key unsafe.Pointer) {
 		return
 	}
 	hash := hash32(key, uintptr(m.keySize))
-	hashmapDelete(m, key, hash, memequal)
+	hashmapDelete(m, key, hash)
 }
 
 // Hashmap with string keys (a common case).
@@ -327,19 +436,24 @@ func hashmapStringHash(s string) uint32 {
 	return hash32(unsafe.Pointer(_s.ptr), uintptr(_s.length))
 }
 
+func hashmapStringPtrHash(sptr unsafe.Pointer, size uintptr) uint32 {
+	_s := *(*_string)(sptr)
+	return hash32(unsafe.Pointer(_s.ptr), uintptr(_s.length))
+}
+
 func hashmapStringSet(m *hashmap, key string, value unsafe.Pointer) {
 	hash := hashmapStringHash(key)
-	hashmapSet(m, unsafe.Pointer(&key), value, hash, hashmapStringEqual)
+	hashmapSet(m, unsafe.Pointer(&key), value, hash)
 }
 
 func hashmapStringGet(m *hashmap, key string, value unsafe.Pointer, valueSize uintptr) bool {
 	hash := hashmapStringHash(key)
-	return hashmapGet(m, unsafe.Pointer(&key), value, valueSize, hash, hashmapStringEqual)
+	return hashmapGet(m, unsafe.Pointer(&key), value, valueSize, hash)
 }
 
 func hashmapStringDelete(m *hashmap, key string) {
 	hash := hashmapStringHash(key)
-	hashmapDelete(m, unsafe.Pointer(&key), hash, hashmapStringEqual)
+	hashmapDelete(m, unsafe.Pointer(&key), hash)
 }
 
 // Hashmap with interface keys (for everything else).
@@ -427,21 +541,26 @@ func hashmapInterfaceHash(itf interface{}) uint32 {
 	}
 }
 
+func hashmapInterfacePtrHash(iptr unsafe.Pointer, size uintptr) uint32 {
+	_i := *(*_interface)(iptr)
+	return hashmapInterfaceHash(_i)
+}
+
 func hashmapInterfaceEqual(x, y unsafe.Pointer, n uintptr) bool {
 	return *(*interface{})(x) == *(*interface{})(y)
 }
 
 func hashmapInterfaceSet(m *hashmap, key interface{}, value unsafe.Pointer) {
 	hash := hashmapInterfaceHash(key)
-	hashmapSet(m, unsafe.Pointer(&key), value, hash, hashmapInterfaceEqual)
+	hashmapSet(m, unsafe.Pointer(&key), value, hash)
 }
 
 func hashmapInterfaceGet(m *hashmap, key interface{}, value unsafe.Pointer, valueSize uintptr) bool {
 	hash := hashmapInterfaceHash(key)
-	return hashmapGet(m, unsafe.Pointer(&key), value, valueSize, hash, hashmapInterfaceEqual)
+	return hashmapGet(m, unsafe.Pointer(&key), value, valueSize, hash)
 }
 
 func hashmapInterfaceDelete(m *hashmap, key interface{}) {
 	hash := hashmapInterfaceHash(key)
-	hashmapDelete(m, unsafe.Pointer(&key), hash, hashmapInterfaceEqual)
+	hashmapDelete(m, unsafe.Pointer(&key), hash)
 }
